Extend room admin api with additional attributes
